# Conflicts:
#   src/backEnd/GameData/State/AccessPermissionsImpl.java
#   src/frontEnd/Skeleton/Canvas.java
#   src/frontEnd/Skeleton/UserTools/TileCommandCenter.java

+++ src/backEnd/GameData/State/TileGrid.java
@@ -24,7 +24,7 @@ public interface TileGrid {
 	 * @param y
 	 * @return
 	 */
-	Tile getTileByCoord(int x, int y);
+	Tile getTileByCoord(Double x, Double y);
 	
 	/**
 	 * @return grid width
